% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtr_functions.R
\name{download_file_vtr}
\alias{download_file_vtr}
\title{Download files from AEC Virtual Tally Room}
\usage{
download_file_vtr(file, election, basedir = NA, verbose = FALSE, cached = TRUE)
}
\arguments{
\item{file}{The name of the file to retrieve}

\item{election}{The AEC's event identifier for the election or the election
year.}

\item{basedir}{(Optional) The URL of the Vitrual Tally Room Page - only
necessary for results of current elections that have not yet been archived
to results.aec.gov.au}

\item{verbose}{(Optional) Be more verbose}

\item{cached}{(Optional) Force using cached files, rather than re-downloading
files from the VTR}
}
\value{
A \code{data.frame} with the CSV data as parsed by \code{read.csv},
  with state/territory files combined where necessary.
}
\description{
Download results files from AEC Virtual Tally Room (VTR) and return a
\code{data.frame} of the data.
}
\details{
The results files on the VTR store the CSV files as data but they are
packages in a number of ways, including as multiple files in a .zip archive,
or as seaprate files for each state and territory (or both). This function
uses the predictable naming structure of the files to get the correct files
and return them as a \code{data.frame} of national results.

The function uses \code{\link{read_cvs_vtr}} internally to read the files,
which is a thin wrapper around \link{read.csv}, but also adds some metadata
to the files as attributes.
}
\examples{
\dontrun{
download_file_vtr("HouseCandidatesDownload",
                  24310,
                  "https://tallyroom.aec.gov.au/Downloads/")
}

}
\seealso{
\code{\link{read_cvs_vtr}} for the function that actually reads the
  CSV files that have been retrieved.
}
