% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{download_mediafeed_api}
\alias{download_mediafeed_api}
\title{Download media feed files via the plumber API}
\usage{
download_mediafeed_api(Server, EventIdentifier, Filetype, Archive = TRUE)
}
\arguments{
\item{Server}{The URL of the server, including protocol and port, with no
trailing slash}

\item{EventIdentifier}{AEC event identifier or election year}

\item{Filetype}{One of \code{Light}, \code{Preload} or \code{Verbose}}

\item{Archive}{If \code{TRUE} (default), use the AEC's media feed archive
site, or \code{FALSE} for the live election results}
}
\value{
A string with the full path of the downloaded file.
}
\description{
The AEC media feed is hosted on an anonymous FTP site, which may be blocked
by some firewalls. This function works in conjunction with a \pkg{plumber}
script, which is included in the \code{plumber} directory in this package.
}
\details{
Simply run the included \code{run.R} file in the \pkg{plumber} directory on a
server that can access the media feed FTP server and use the resulting server
address and port with this function to download the latest zipfile of the
appropriate type from the media feed.

The \code{api.R} file in the \pkg{plumber} directory runs
\code{\link{download_mediafeed_file}} with the correct parameters and sends
the resulting file, which is then downloaded by this function.

Note that the function does not give any feedback when it is downloading the
file (which might take a while as some of them are large), but does do a bit
of error checking when the download is done.
}
\examples{
\dontrun{
download_mediafeed_api("http://127.0.0.1:3379", 2022, "Preload", Archive = FALSE)}
}
