% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{download_mediafeed_file}
\alias{download_mediafeed_file}
\title{Download file from AEC media feed FTP site}
\usage{
download_mediafeed_file(
  EventIdentifier,
  Filetype,
  Detail = "Detailed",
  Dest = tempdir(),
  Archive = TRUE,
  Download = TRUE
)
}
\arguments{
\item{EventIdentifier}{AEC event identifier or election year}

\item{Filetype}{One of \code{Light}, \code{Preload}, \code{PollingDistricts}
or \code{Verbose}}

\item{Detail}{One of either \code{Detailed} (default) or \code{Standard}}

\item{Dest}{Destination path for the downloaded file (will be created if it
does not exist) and defaults to \code{tempdir()}}

\item{Archive}{If \code{TRUE} (default), use the AEC's media feed archive
site, or \code{FALSE} for the live election results}

\item{Download}{(\code{LOGICAL}) If \code{TRUE} (default) download the file
and return the path on the local disk. If \code{FALSE}, don't download the
file but return the URL where it located.}
}
\value{
A file path of the downloaded file or the URL of the file
}
\description{
Downloads a specific files from the AEC's media feed FTP site and returns the
filename and path of the file.
}
\examples{
\dontrun{
download_mediafeed_file(2022, "Preload", "Detailed", "/tmp")
}

}
\seealso{
\link{read_mediafeed_xml} for reading the resulting file.
}
