% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{fetch_mediafeed_results}
\alias{fetch_mediafeed_results}
\title{Fetch election results media feed}
\usage{
fetch_mediafeed_results(election, api = FALSE, archive = FALSE)
}
\arguments{
\item{election}{AEC EventIdentifier or election year}

\item{api}{If \code{FALSE} (default), download the file from the AEC's FTP
site, otherwise, the URL "including port and trailing slash" of the server
running the media feed API.}

\item{archive}{If true, use the media feed archive site, if \code{FALSE}
(default), use the live results media feed (only available during an
election).}
}
\value{
An XML media feed object.
}
\description{
Fetch the current media feed results xml for the nominated election.
}
\details{
This is simply a shortcut for \code{\link{download_mediafeed_file}} and
\code{\link{read_mediafeed_xml}}.
}
\examples{
\dontrun{resuls_xml <- fetch_mediafeed_results(2022)}
}
