% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Fill}
\alias{Fill}
\title{Fill down}
\source{
\url{https://stackoverflow.com/questions/9514504/add-missing-value-in-column-with-value-from-row-above}
}
\usage{
Fill(x)
}
\arguments{
\item{x}{A vector or \code{data.frame} column.}
}
\value{
A vector or \code{data.frame} column.
}
\description{
Internal convenience fill down function to avoid including all of \code{tidyr}.
}
