% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{get_mediafeed_analysis}
\alias{get_mediafeed_analysis}
\title{Get national media feed analysis}
\usage{
get_mediafeed_analysis(xml)
}
\arguments{
\item{xml}{An XML media feed object.}
}
\value{
A \code{data.frame} with seven variables: \code{CandidateType} (one
  of \code{PartyGroup}, \code{AmalgamatedGhostGroups}, \code{Formal},
  \code{Informal} or \code{Total}), \code{PartyGroupId},
  \code{PartyGroupCode}, \code{PartyGroupNm}, \code{Percentage}, \code{Swing}
  and \code{Votes}.
}
\description{
Get the AEC's aggregate results by party group nationally.
}
\examples{
results_xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Verbose", Archive = TRUE))
get_mediafeed_analysis(results_xml)

}
\seealso{
\code{\link{get_mediafeed_analysis_states}} for the same results
  disaggregated by state.
}
