% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{get_mediafeed_analysis_states}
\alias{get_mediafeed_analysis_states}
\title{Get media feed analysis by state}
\usage{
get_mediafeed_analysis_states(xml)
}
\arguments{
\item{xml}{An XML media feed object}
}
\value{
A \code{data.frame} with eight variables: \code{StateAb},
  \code{CandidateType} (one of \code{PartyGroup},
  \code{AmalgamatedGhostGroups}, \code{Formal}, \code{Informal} or
  \code{Total}), \code{PartyGroupId}, \code{PartyGroupCode},
  \code{PartyGroupNm}, \code{Percentage}, \code{Swing} and \code{Votes}.
}
\description{
Get the AEC's aggregate results by party group by state.
}
\examples{
results_xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Verbose", Archive = TRUE))
get_mediafeed_analysis_states(results_xml)

}
\seealso{
\code{\link{get_mediafeed_analysis}} for the same results as a
  national total.
}
