% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{get_mediafeed_candidates}
\alias{get_mediafeed_candidates}
\title{DEPRECATECD: Get House election candidate details from results media feed}
\usage{
get_mediafeed_candidates(xml, party = FALSE)
}
\arguments{
\item{xml}{A pointer to an XML media feed object}

\item{party}{(\code{LOGICAL}) Fetch party ID (defaults to \code{FALSE}
becuase it is very slow.)}
}
\value{
A \code{data.frame} with 9 variables (10 if \code{party} is
  \code{TRUE}).
}
\description{
List all of the candidates in the results media feed, along with their
CandidateId and PartyId, mostly for merging with results by CandidateId.
}
\details{
Note that the list of candidates in the results feed is different from the
candidates in the candidates file in the preload. The candidates file in the
preload includes the candidate's gender, but does not include 'ghost'
candidates (candidates who ran in the last election but not this election,
included for calculating swings). The results file contains current and ghost
candidates, but not some other details like gender.

The function can also optionally return the party ID for each candidate,
however the data extraction process can be quite slow because it needs to
search the XML separately for the party affiliation of each candidate. As
such, if you do use this you should plan to run this once and save it, rather
than running it each time you do the analysis. But even better, you should
have already got this from the preload file.
}
\examples{
\dontrun{
results_xml <- read_mediafeed_xml(get_mediafeed_file(2022, "Verbose", Archive = TRUE))
get_mediafeed_fp_by_pps(results_xml)}
}
