% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{get_mediafeed_division_details}
\alias{get_mediafeed_division_details}
\title{Get Division Details from Media Feed}
\usage{
get_mediafeed_division_details(DivisionID, xml)
}
\arguments{
\item{DivisionID}{A division ID in either character or interger format.}

\item{xml}{A pointer to an XML media feed object.}
}
\value{
A \code{data.frame}.
}
\description{
Get Division Details from Media Feed
}
\examples{
\dontrun{
library(purrr)
xml <- read_mediafeed_xml(get_mediafeed_file(2022, "Verbose", Archive = FALSE))
map_dfr(get_mediafeed_divisionids(xml), get_mediafeed_division_details, xml)
}

}
