% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{get_mediafeed_division_status}
\alias{get_mediafeed_division_status}
\title{Get House division count status from media feed}
\usage{
get_mediafeed_division_status(DivisionID, xml)
}
\arguments{
\item{DivisionID}{A division ID in either character or integer format.}

\item{xml}{A pointer to an XML media feed object.}
}
\value{
A \code{data.frame}.
}
\description{
Return details of the specified division from the media feed file as a
single-row \code{data.frame}.
}
\examples{
\dontrun{
library(purrr)
xml <- read_mediafeed_xml(get_mediafeed_file(2022, "Verbose", Archive = FALSE))
map_dfr(get_mediafeed_divisionids(xml), get_mediafeed_division_status, xml)
}

}
