% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{get_mediafeed_divisionids}
\alias{get_mediafeed_divisionids}
\title{Get the list of Division IDs from the media feed}
\usage{
get_mediafeed_divisionids(xml)
}
\arguments{
\item{xml}{A pointer to an XML media feed object.}
}
\value{
A vector of division IDs.
}
\description{
List House division IDs from media feed file.
}
\examples{
\dontrun{
xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Verbose", Archive = FALSE))
get_mediafeed_divisionids(xml)
}
}
