% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{get_mediafeed_fp_by_pps}
\alias{get_mediafeed_fp_by_pps}
\title{Get first preference votes by polling place}
\usage{
get_mediafeed_fp_by_pps(xml)
}
\arguments{
\item{xml}{A pointer to an XML media feed object}
}
\value{
A \code{data.frame} with seven columns: \code{PollingPlaceId},
  \code{CandidateType} (one of either \code{Candidate} or \code{Ghost}),
  \code{CandidateID}, \code{FP.Historic}, \code{FP.Percentage},
  \code{FP.Swing} and \code{FP.Votes}.
}
\description{
Extract the first preference votes by candidate by polling place from the
media feed file, including the historic, percentage and swing.
}
\details{
This function is designed to be as quick as minimal as possible whilst still
including robust error checking and feedback.

The actual results are prefixed with "FP." in order to easily distintinguish
them from two-candidate preferred and two-party preferred votes.

It assumes that you already have the candidate, divisions and polling place
details elsewhere, and will join those by PollingPlaceId and CandidateID. In
particular, this function does not include DivisionID or PartyID.

This function is quite vociferous because 1) it can be slow, so it tells you
what it's doing, and 2) occasionally the media feed changes and this helps
you determine where the function has made assumptions that are no longer
true.
}
\examples{
\dontrun{
xml <- read_mediafeed_xml(get_mediafeed_file(2022, "Verbose", Archive = TRUE))
get_mediafeed_fp_by_pps(xml)}

}
