% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_mediafeed_metadata}
\alias{get_mediafeed_metadata}
\title{Print media feed metadata}
\usage{
get_mediafeed_metadata(xml, short = TRUE)
}
\arguments{
\item{xml}{A pointer to an XML media feed object.}

\item{short}{\code{TRUE} to display just the most important items (default)
or \code{FALSE} to display everything.}
}
\value{
A named vector.
}
\description{
Display metadata from the media feed XML file.
}
\details{
Display some metadata from the media feed XML file, including when it was
updated, its verbosity and event identifier.

This should be relatively quick because it only accesses information at the
very top of the file (although there is inevitably some delay in parsing even
the first bit of such a large file).

The time stamp returned in \code{Created} can be converted with
\code{"\%FT\%T"} as a format string.
}
\examples{
\dontrun{
results_xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Verbose", Archive = FALSE))
get_mediafeed_metadata(results_xml)}

}
