% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_preload_functions.R
\name{get_mediafeed_preload_candidates}
\alias{get_mediafeed_preload_candidates}
\title{Get candidates from preload media feed file}
\usage{
get_mediafeed_preload_candidates(xml)
}
\arguments{
\item{xml}{A pointer to an XML preload media feed object}
}
\value{
A \code{data.frame} with 20 variables.
}
\description{
Get candidates from preload media feed file
}
\examples{
preload_xml <- read_mediafeed_xml(download_mediafeed_file(2022,  "Preload", Archive = TRUE),
                                  "results")
get_mediafeed_preload_candidates(preload_xml)
}
