% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_preload_functions.R
\name{get_mediafeed_preload_candidates_sen}
\alias{get_mediafeed_preload_candidates_sen}
\title{Get Senate candidates from preload media feed file}
\usage{
get_mediafeed_preload_candidates_sen(xml)
}
\arguments{
\item{xml}{A pointer to an XML preload media feed object.}
}
\value{
A \code{data.frame}.
}
\description{
Get Senate candidates from preload media feed file
}
