% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_preload_functions.R
\name{get_mediafeed_preload_divs}
\alias{get_mediafeed_preload_divs}
\title{Get divisions from preload media feed file}
\usage{
get_mediafeed_preload_divs(xml)
}
\arguments{
\item{xml}{A pointer to an XML preload media feed object.}
}
\value{
a \code{data.frame} with seven variables: \code{DivisionID},
  \code{DivisionShortCode} (a four-character code for the division),
  \code{DivisionNm}, \code{StateAb}, \code{Enrolment.CloseOfRolls},
  \code{Enrolment.Historic} and \code{Enrolment}
}
\description{
Get divisions and enrolments from the preload media feed XML file.
}
\details{
This function extracts the divisions and their associated identifiers
(including \code{DivisionShortCode}, which only seems to be used in the media
feed and nowhere else).

Note that \code{\link{get_mediafeed_preload_pps}} includes \code{DivisionID}
and \code{DivisionShortName} but not \code{DivisonNm} - this is the only of
the preload functions where division name is available, so it makes sense to
join this table to the polling place table by \code{DivisionID}.

This function also outputs various enrolment numbers for the division.
\code{EnrolmentHistoric} can be used for enrolment swing calculations, but
it's not immediately clear whether \code{Enrolment} or
\code{Enrolment.CloseOfRolls} is the most correct number for calculating
turnout.
}
\examples{
preload_xml <- read_mediafeed_xml(download_mediafeed_file(2022,
                                                          "Preload",
                                                          Archive = TRUE),
                                  "results")
 get_mediafeed_preload_divs(preload_xml)
}
