% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_preload_functions.R
\name{get_mediafeed_preload_gender}
\alias{get_mediafeed_preload_gender}
\title{Get candidate gender from preload media feed}
\usage{
get_mediafeed_preload_gender(xml, chamber)
}
\arguments{
\item{xml}{A pointer to an XML media feed object.}

\item{chamber}{One of either \code{House} or \code{Senate}.}
}
\value{
A \code{data.frame} with three columns: \code{CandidateId},
  \code{Gender} ("male" or "female"), and \code{EventId}
}
\description{
Create a \code{data.frame} of candidate IDs and gender from the candidates
file in the media feed preload file.
}
\details{
Note that the gender information is only found in the
\code{eml-230-candidates-(EventId).xml} file, which is in the \code{preload}
zip file. If the correct XML node identifying this file
(\code{d1:CandidateList}) is not found, the function will halt with an error.

The easiest way to get the file is to call \code{\link{read_mediafeed_xml}}
with "candidates" as the \code{filename} argument,
}
\examples{
\dontrun{
file <- download_mediafeed_file(2022, "Preload", "Detailed")
xml <- read_mediafeed_xml(file, "candidates")
get_mediafeed_preload_gender(xml, chamber = "senate")}

}
