% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_preload_functions.R
\name{get_mediafeed_preload_parties}
\alias{get_mediafeed_preload_parties}
\title{Get parties from preload media feed file}
\usage{
get_mediafeed_preload_parties(xml)
}
\arguments{
\item{xml}{A pointer to an XML preload media feed object.}
}
\value{
A \code{data.frame} with three variables: \code{PartyGroupId}
  (\code{INT}), \code{PartyGroupShortCode} and \code{PartyGroupName}.
}
\description{
Extract a table of parties, party IDs and shortcodes from the preload XML
file.
}
\examples{
\dontrun{
preload_xml <- read_mediafeed_xml(download_mediafeed_file(2022,
                                                          "Preload",
                                                          Archive = TRUE),
                                  "results")
get_mediafeed_preload_parties(preload_xml)}
}
