% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_preload_functions.R
\name{get_mediafeed_preload_partygroups}
\alias{get_mediafeed_preload_partygroups}
\title{Get party groups from preload media feed file}
\usage{
get_mediafeed_preload_partygroups(xml)
}
\arguments{
\item{xml}{A pointer to an XML preload media feed object.}
}
\value{
A \code{data.frame} with three variables: \code{PartyGroupID}
  (\code{INT}), \code{PartyGroupShortCode} and \code{PartyGroupName}.
}
\description{
Extract a table of party group, IDs and shortcodes from the preload XML file.
}
\details{
Note that party groups are not the same as the parties listed in the
candidate extract. I'm not sure that there is any way to determine which
parties are members of which groups from just the preload files.
}
\examples{
\dontrun{
preload_xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Preload", Archive = TRUE),
                                  "results")
get_mediafeed_preload_parties(preload_xml)}
}
