% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_preload_functions.R
\name{get_mediafeed_preload_pps}
\alias{get_mediafeed_preload_pps}
\title{Get polling places from preload media feed file}
\usage{
get_mediafeed_preload_pps(xml)
}
\arguments{
\item{xml}{A pointer to an XML preload media feed object.}
}
\value{
A \code{data.frame} with five variables: \code{DivisionId},
  \code{DivisionShortCode}, \code{PollingPlaceId}, \code{PollingPlaceName}
  and \code{PollingPlaceClassification}.
}
\description{
Get polling places and their associated divisions from the preload media feed
XML file.
}
\details{
Note that the division name is not included is the output of this function,
only the \code{DivisionId} and \code{DivisionShortCode}. To get
\code{DivisionNm}, join this file with the output of
\code{\link{get_mediafeed_preload_divs}} on \code{DivisionId}.

Most of the polling places to not have a \code{PollingPlaceClassification}
(that is, they are \code{NA}). As of the 2022 federal election, the
classifications are \code{PrePollVotingCentre}, \code{SpecialHospital},
\code{PrisonMobile} and \code{RemoteMobile}.

The blind and low vision telephone voting (and for 2022 the COVID telephone
voting) are classed as \code{PrePollVotingCentre} polling places, but are
only identifiable by the \code{PollingPlaceName}.
}
\examples{
\dontrun{
preload_xml <- read_mediafeed_xml(download_mediafeed_file(2022,
                                                          "Preload",
                                                          Archive = TRUE),
                                  "results")
get_mediafeed_preload_pps(preload_xml)}
}
