% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_preload_functions.R
\name{get_mediafeed_preload_pps_cand}
\alias{get_mediafeed_preload_pps_cand}
\title{Get candidates and polling places from preload media feed}
\usage{
get_mediafeed_preload_pps_cand(xml)
}
\arguments{
\item{xml}{A pointer to an XML preload media feed object}
}
\value{
A \code{data.frame} with the following variables: \code{StateAb},
  \code{DivisionId}, \code{DivisionShortCode}, \code{DivisionNm},
  \code{PollingPlaceId}, \code{PollingPlaceNm}, \code{CandidateId},
  \code{CandidateType}, \code{IsGhost}, \code{Historic}, \code{Percentage},
  \code{Swing}, \code{Votes}.
}
\description{
Get a \code{data.frame} with candidates by polling place in the same format
as the media feed results file.
}
\details{
This function is similar to \code{\link{get_mediafeed_preload_pps}} and
\code{\link{get_mediafeed_preload_candidates}}, but returns a table of
candidates by polling place.

The usefulness of this function is that it returns the table of candidates
and polling places in exactly the same format and order as
\code{\link{get_mediafeed_votes_pps_fp}}, which makes it trivial to merge
with the live results file (you don't even have to join it, you could
\code{\link{cbind}} it, but you should probably also check that
\code{PollingPlaceId} and \code{CandidateId} are equal).
}
\examples{
preload_xml <- read_mediafeed_xml(download_mediafeed_file(2022,  "Preload", Archive = TRUE),
                                  "results")
get_mediafeed_preload_pps_cand(preload_xml)
}
