% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_preload_functions.R
\name{get_mediafeed_preload_votes_type}
\alias{get_mediafeed_preload_votes_type}
\title{Get vote type table from preload media feed file}
\usage{
get_mediafeed_preload_votes_type(xml)
}
\arguments{
\item{xml}{A pointer to an XML preload media feed object}
}
\value{
a \code{data.frame}. See \code{\link{get_mediafeed_votes_type}} for
  details.
}
\description{
Extract votes by type (Ordinary, Absent, Provisional, PrePoll or Postal) from
the media feed.
}
\details{
This just calls \code{\link{get_mediafeed_votes_type}} on the preload file
for first preference votes, and returns historic votes by type.
}
\examples{
preload_xml <- read_mediafeed_xml(download_mediafeed_file(2022,  "Preload", Archive = TRUE),
                                  "results")
get_mediafeed_preload_votes_type(preload_xml)
}
