% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senate_functions.R
\name{get_mediafeed_senate_div}
\alias{get_mediafeed_senate_div}
\title{Get Senate votes by division}
\usage{
get_mediafeed_senate_div(xml)
}
\arguments{
\item{xml}{A pointer to an XML media feed object.}
}
\value{
A \code{data.frame}.
}
\description{
Get Senate votes by division
}
