% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senate_functions.R
\name{get_mediafeed_senate_quotas}
\alias{get_mediafeed_senate_quotas}
\title{Get Senate quotas by state}
\usage{
get_mediafeed_senate_quotas(xml)
}
\arguments{
\item{xml}{A pointer to an XML media feed object.}
}
\value{
A \code{data.frame}.
}
\description{
Get Senate quotas by state
}
