% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mediafeed_votes_by_pps.R
\name{get_mediafeed_votes_by_pps}
\alias{get_mediafeed_votes_by_pps}
\title{Extract votes from the media feed by polling place}
\description{
Extract votes from the media feed by polling place
}
