% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mediafeed_votes_div.R
\name{get_mediafeed_votes_div}
\alias{get_mediafeed_votes_div}
\title{Get votes by division}
\usage{
get_mediafeed_votes_div(xml, type)
}
\arguments{
\item{xml}{A pointer to an XML media feed object}

\item{type}{One of either \code{fp} (first preference), \code{tcp}
(two-candidate preferred), \code{tpp} (two party preferred) or \code{total}
(total first preference votes).}
}
\value{
A \code{data.frame} with a \code{DivisionId} column and at least
  current, historic, percentage, swing and matched historic votes.
}
\description{
Generic function for downloading various vote types (first preference,
two-candidate preferred, two-party preferred or total votes) by polling
place.
}
\details{
Note that this iteration of this function does not have any real error
checking, and may stop working if it receives unexpected input. The data
extraction process itself should be relatively robust, but the type
conversions in the function are hardcoded and may fail.
}
\examples{
\dontrun{
results_xml <- read_mediafeed_xml(get_mediafeed_file(2022, "Verbose", Archive = TRUE))
get_mediafeed_votes_div(results_xml, "fp")}

}
