% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mediafeed_votes_pps.R
\name{get_mediafeed_votes_pps_tcp}
\alias{get_mediafeed_votes_pps_tcp}
\title{Get two-candidate preferred votes by polling place}
\usage{
get_mediafeed_votes_pps_tcp(xml)
}
\arguments{
\item{xml}{A pointer to an XML media feed object}
}
\description{
Get two-candidate preferred votes by polling place
}
