% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mediafeed_votes_by_pps.R
\name{get_mediafeed_votes_pps_total}
\alias{get_mediafeed_votes_pps_total}
\title{Get total votes by polling place}
\usage{
get_mediafeed_votes_pps_total(xml)
}
\arguments{
\item{xml}{A pointer to an XML media feed object}
}
\description{
Get total votes by polling place
}
