% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senate_functions.R
\name{get_mediafeed_votes_sen}
\alias{get_mediafeed_votes_sen}
\title{Get Senate votes by state}
\usage{
get_mediafeed_votes_sen(xml)
}
\arguments{
\item{xml}{A pointer to an XML media feed object.}
}
\value{
A \code{data.frame} with 13 variables: \code{StateAb},
  \code{CandidateType} (can be one of either \code{Candidate},
  \code{TicketVotes}, \code{Unapportioned} or \code{UngroupedCandidate}),
  \code{Ticket}, \code{GroupId}, \code{CandidateId}, \code{PartyId},
  \code{IsIndependent}, \code{Historic} (votes), \code{Percentage},
  \code{Swing}, \code{QuotaProportion} and \code{Votes}.
}
\description{
Get the Senate first preference votes from the media feed XML file by state
and territory.
}
\details{
Similar to the equivalent House functions, this function attempts to return a
relatively minimalist table of results that is designed to be merged with the
output of \code{\link{get_mediafeed_preload_candidates_sen}}.

There are three possible values for \code{CandidateType}:

\itemize{

\item{\code{Candidate}}{A candidate in a group}

\item{\code{TicketVotes}}{The total number of votes above the line for that
group}

\item{\code{Unapportioned}}{According to the
\href{https://www.aec.gov.au/footer/glossary.htm#u}{AEC's definitions},
unapportioned votes are "are votes that have been allocated to a party or
group by one count but have not yet been allocated to individual candidates
within that party or group. As the counting progresses, these votes will be
distributed to the individual candidates."}

\item{\code{UngroupedCandidate}}{A candidate not in a Senate group (that is,
without a box above the line).}

}

The media feed also provides a \code{GroupTotal} result, but for the sake of
convenience that is not included in the output of this function because it
would require filtering out before aggregating votes.
}
\examples{
results_xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Verbose", Archive = TRUE))
get_mediafeed_senate_quotas(results_xml)

}
