% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{get_mediafeed_votes_type}
\alias{get_mediafeed_votes_type}
\title{Get votes by type from the media feed}
\usage{
get_mediafeed_votes_type(xml, count = "fp")
}
\arguments{
\item{xml}{A pointer to an XML media feed object.}

\item{count}{Currently \code{fp} for first preferences or \code{tcp} for
two-candidate preferred.}
}
\value{
a \code{data.frame} with seven variables: \code{CandidateId},
  \code{DivisionId}, \code{Type} (\code{Ordinary}, \code{Absent},
  \code{Provisional}, \code{PrePoll}, \code{Postal}), \code{Historic},
  \code{Percentage}, \code{Swing} and \code{Votes}.
}
\description{
Extract votes by type (Ordinary, Absent, Provisional, PrePoll or Postal) from
the media feed.
}
\details{
All of the vote types except \code{Ordinary} votes are types of declaration
votes.

It is important to note that the \code{PrePoll} votes returned by this
function are pre-poll declaration votes, which are pre-poll votes cast
outside the elector's division. Pre-poll ordinary votes are what most people
mean by "pre-poll votes". These can be found through the
\code{\link{get_mediafeed_preload_pps}} function, and have a
\code{PollingPlaceClassification} of \code{PrePollVotingCentre}.

Note that for maximum extract speed the data is extracted in the tabular
format that it appears in the media feed XML file, which is probably not
optimal for most analysis. For best results, the resulting \code{data.frame}
should probably be pivoted into a \code{tidyr} "long" format.
}
\examples{
\dontrun{
results_xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Verbose", Archive = TRUE))
get_mediafeed_votes_type(results_xml, count = "fp")}

}
