% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_functions.R
\name{mediafeed_display}
\alias{mediafeed_display}
\title{Display current election results}
\usage{
mediafeed_display(xml, preload_candidates)
}
\arguments{
\item{xml}{A pointer to an XML media feed object.}

\item{preload_candidates}{A \code{data.frame} of candidate details obtained
from \code{\link{get_mediafeed_preload_candidates}}.}
}
\value{
A list with tables for each division. Note that the names of each of
  the division results list items are the division ids as a character string.
}
\description{
Return a list of election results tables with current results suitable for
displaying in an application or function.
}
\examples{
\dontrun{
preload_xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Preload", Archive = TRUE),
                                  "results")

mf_cand <- get_mediafeed_preload_candidates(preload_xml)

results_xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Verbose", Archive = TRUE))
mediafeed_display(results_xml, mf_cand)}
}
