% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_preload_functions.R
\name{mediafeed_preload_functions}
\alias{mediafeed_preload_functions}
\title{Extract election data from media feed preload files}
\description{
This is a collection of functions that are designed to be run on the preload
media feed files (in fact they check that they're running on the preload
files because at least one of them depends on a set of nodes that is only in
the preload files - the \code{//d1:Analysis} nodes).
}
\details{
The intention is that these functions are run on the preload file before the
election (or before any other results are analysed) to build tables that can
be joined against the results files. Extracting this data is (relatively)
slow, and extracting the results is as quick as it can be.

The functions included in this family are:

\itemize{

\item{\code{\link{get_mediafeed_preload_parties}}}{Get parties from preload
media feed file}

\item{\code{\link{get_mediafeed_preload_candidates}}}{Get candidates from
preload media feed file}

\item{\code{\link{get_mediafeed_preload_pps}}}{Get polling places from
preload media feed file}

\item{\code{\link{get_mediafeed_preload_divs}}}{Get divisions from preload
media feed file}

\item{\code{\link{get_mediafeed_preload_votes_type}}}{Get vote type table
from preload media feed file}

\item{\code{\link{get_mediafeed_preload_gender}}}{Get candidate gender from
preload media feed}

}
}
