% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{merge_div_candidates}
\alias{merge_div_candidates}
\title{Merge candidate details to division results}
\usage{
merge_div_candidates(preload_cand, xml, type)
}
\arguments{
\item{preload_cand}{A \code{data.frame} generated by
\code{\link{get_mediafeed_preload_candidates}}.}

\item{xml}{A pointer to an XML media feed object.}

\item{type}{Either "fp" (first preference) or "tcp" (two-candidate
preferred).}
}
\value{
A \code{data.frame}
}
\description{
Add candidate and division details to division-level results by
\code{CandidateId} and \code{DivisionId}.
}
\details{
This function runs \code{\link{get_mediafeed_votes_div}} and merges the
results with the output of \code{\link{get_mediafeed_preload_candidates}},
which contains all of the candidate names (and details) and division names.
In practice, this is a much more useful function for analysis of results than
\code{\link{get_mediafeed_votes_div}} by itself.
}
\examples{
\dontrun{
preload_xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Preload", Archive = TRUE),
                                  "results")

mf_cand <- get_mediafeed_preload_candidates(preload_xml)

results_xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Verbose", Archive = TRUE))

merge_div_candidates(mf_cand, results_xml, "fp")}

}
\seealso{
\code{\link{get_mediafeed_votes_div}} for the details of the
  function that actually fetches the results.
}
