% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{read_mediafeed_xml}
\alias{read_mediafeed_xml}
\title{Read AEC media feed XML file}
\usage{
read_mediafeed_xml(path, filename = NA)
}
\arguments{
\item{path}{Full path to zip file or XML file}

\item{filename}{In the case of the preload zip file, one of \code{results},
\code{pollingdistricts}, \code{event} or \code{candidates}}
}
\value{
An \code{XML2} library XML pointer
}
\description{
Read the media feed XML file and return an \code{xml2} pointer, unzipping the file if necessary.
}
\details{
This is either a very thin wrapper around \code{xml2::read_xml()} or an easy
way to access a media feed XML file that is contained in a zip file, such as
one downloaded by \link{get_mediafeed_file}.
}
\examples{
\dontrun{
# Read a file that you have downloaded yourself
read_mediafeed_xml("/tmp/aec-mediafeed-Detailed-Preload-27966-20220503123540.zip", "results")

# Read a file that has been downloaded with read_mediafeed_xml()
read_mediafeed_xml(download_mediafeed_file(2022, "Preload", "Detailed", "/tmp"))
}

}
\seealso{
\link{get_mediafeed_file} for downloading the correct file from the
  AEC's FTP site.
}
