% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{utility_functions}
\alias{utility_functions}
\title{Retrieve the xml files from the media feed}
\description{
These functions are used to retrieve the files from the AEC's FTP server
and make the contained XML available to be used with the package's
other functions.
}
