% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausvotes_functions.R
\name{vtr_identifier}
\alias{vtr_identifier}
\title{VTR Identifier as Year or EventId}
\usage{
vtr_identifier(identifier, as)
}
\arguments{
\item{identifier}{One of either an election year or an AEC event identifier}

\item{as}{One of either \code{year} or \code{event}}
}
\value{
Either the election year or event identifier as a numeric.
}
\description{
Convert an event identifier or an election year into either an event identifier
or an election year.
}
\details{
The Australian Electoral Commission (AEC) identifies each Australian federal
election and by-election with a (so far) five digit event identifier. This is
used in the labelling of the media feed files and in the file paths of the
files on the virtual tally room, but is not easy to remember (or determine).

This function allows other functions that require an election identifier to
accept either a year or an election identifier. For convenience it also
works the other way around, in case you need that.
}
\examples{
vtr_identifier(27966, "year")

vtr_identifier(2022, "event")
}
