% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausvotes_functions.R
\name{vtr_ids}
\alias{vtr_ids}
\title{VTR Ids}
\usage{
vtr_ids()
}
\value{
A \code{data.frame} with one column of \code{EventId} and
        one column \code{ElectionYear}.
}
\description{
A \code{data.frame} with EventIds and ElectionYear for use in other functions.
}
\details{
This is designed to be the one place with a mapping between election years
and AEC election identifiers.
}
