% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_preload_functions.R
\name{get_mediafeed_preload_candidates}
\alias{get_mediafeed_preload_candidates}
\title{Get candidates from preload media feed file}
\usage{
get_mediafeed_preload_candidates(xml)
}
\arguments{
\item{xml}{A pointer to an XML preload media feed object}
}
\value{
A \code{data.frame} with 21 variables: \code{DivisionId},
  \code{DivisionShortCode}, \code{DivisionNm}, \code{CandidateType}
  (\code{Candidate} or \code{Ghost}), \code{IsGhost} (\code{logical}),
  \code{CandidateId}, \code{CandidateNm}, \code{BallotPosition},
  \codde{IsIndependent} (\code{logical}), \code{PartyId}, \code{PartyCode},
  \code{PartyNm}, \code{Elected} (\code{logical}), \code{ElectedHistoric}
  (\code{logical}), \code{Incumbent} (\code{logical}),
  \code{IncumbentNotional} (\code{logical}), \code{FP.Votes},
  \code{FP.Percentage}, \code{FP.Swing}, \code{FP.Historic},
  \code{FP.MatchedHistoric}.
}
\description{
Get candidates from preload media feed file
}
\examples{
preload_xml <- read_mediafeed_xml(download_mediafeed_file(2022,  "Preload", Archive = TRUE),
                                  "results")
get_mediafeed_preload_candidates(preload_xml)
}
