% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senate_functions.R
\name{get_mediafeed_senate_quotas}
\alias{get_mediafeed_senate_quotas}
\title{Get Senate quotas by state}
\usage{
get_mediafeed_senate_quotas(xml)
}
\arguments{
\item{xml}{A pointer to an XML media feed object.}
}
\value{
A \code{data.frame} with four variables: \code{StateAb},
  \code{Quota}, \code{QuotaProvisional} (\code{logical}) and
  \code{NumberOfPositions}.
}
\description{
Get AEC calculated Senate quotes and vacancies by jurisdiction from the media
feed file.
}
\details{
This returns the current quota and number of vacancies from the media feed
file for the Senate elections.

Note that if you are planning on using \code{\link{get_mediafeed_votes_sen}}
that provides the votes as quotas through the \code{QuotaProportion}
variable.

This can also be used on the preload XML file, although it will return all of
the \code{Quota} variables as 1. This is mostly useful for functions that
need the \code{NumberOfPositions} variables.
}
\examples{
results_xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Verbose", Archive = TRUE))
get_mediafeed_senate_quotas(results_xml)

}
