% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafeed_functions.R
\name{get_mediafeed_votes_pps}
\alias{get_mediafeed_votes_pps}
\title{Get votes by polling place for division from media feed}
\usage{
get_mediafeed_votes_pps(DivisionID, xml)
}
\arguments{
\item{DivisionID}{A division ID in either character or interger format.}

\item{xml}{A pointer to an XML media feed object.}
}
\value{
a \code{data.frame} with seven columns: \code{DivisionID},
  \code{PollingPlaceID}, \code{CandidateID}, \code{CandidateType},
  \code{Which} and \code{EventIdentifier}.
}
\description{
Extract the votes by candidate and polling place for a division from the
media feed.
}
\details{
Note: The different vote types are \code{Candidate} (normal candidate votes),
\code{Ghost} (Candidates from previous elections not running in this
election, included for swing purpose), \code{Formal} (total formal votes),
\code{Informal}, \code{TotalVotes} and \code{TCP Votes} (two candidate
preferred votes).
}
\examples{
\dontrun{
library(purrr)
xml <- read_mediafeed_xml(get_mediafeed_file(2022, "Verbose", Archive = FALSE))
map_dfr(get_mediafeed_divisionids(xml), get_mediafeed_votes_pps, xml)
}
}
