% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mediafeed_votes_pps.R
\name{get_mediafeed_votes_pps_tcp}
\alias{get_mediafeed_votes_pps_tcp}
\title{Get two-candidate preferred votes by polling place}
\usage{
get_mediafeed_votes_pps_tcp(xml)
}
\arguments{
\item{xml}{A pointer to an XML media feed object}
}
\value{
A \code{data.frame} with the following variables:
  \code{PollingPlaceID}, \code{CandidateID}, \code{TCP.Historic},
  \code{TCP.Percentage}, \code{TCP.Swing}, \code{TCP.Votes}.
}
\description{
Get two-candidate preferred votes by polling place
}
\examples{
xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Verbose", Archive = TRUE))
get_mediafeed_votes_pps_tcp(xml)

}
