% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mediafeed_votes_pps.R
\name{get_mediafeed_votes_pps_total}
\alias{get_mediafeed_votes_pps_total}
\title{Get total votes by polling place}
\usage{
get_mediafeed_votes_pps_total(xml)
}
\arguments{
\item{xml}{A pointer to an XML media feed object}
}
\value{
A \code{data.frame} with six variables: \code{PollingPlaceID},
  \code{VoteTpe} (one of either \code{Formal}, \code{Informal} or
  \code{Total}), \code{Historic}, \code{Percentage}, \code{Swing} and
  \code{Votes}.
}
\description{
Get total votes by polling place
}
\examples{
xml <- read_mediafeed_xml(download_mediafeed_file(2022, "Verbose", Archive = TRUE))
get_mediafeed_votes_pps_total(xml)

}
